feat(magic-link): add allowedAttempts options

+++ packages/better-auth/src/plugins/magic-link/index.ts
@@ -26,6 +26,12 @@ export interface MagicLinkOptions {
 	 */
 	expiresIn?: number | undefined;
 	/**
+	 * Allowed attempts for verifying the magic link token.
+	 * Note: Passing Infinity will allow unlimited attempts.
+	 * @default 1
+	 */
+	allowedAttempts?: number;
+	/**
 	 * Send magic link implementation.
 	 */
 	sendMagicLink: (
@@ -135,6 +141,7 @@ const magicLinkVerifyQuerySchema = z.object({
 export const magicLink = (options: MagicLinkOptions) => {
 	const opts = {
 		storeToken: "plain",
+		allowedAttempts: 1,
 		...options,
 	} satisfies MagicLinkOptions;
 
@@ -209,7 +216,7 @@ export const magicLink = (options: MagicLinkOptions) => {
 					const storedToken = await storeToken(ctx, verificationToken);
 					await ctx.context.internalAdapter.createVerificationValue({
 						identifier: storedToken,
-						value: JSON.stringify({ email, name: ctx.body.name }),
+						value: JSON.stringify({ email, name: ctx.body.name, attempt: 0 }),
 						expiresAt: new Date(Date.now() + (opts.expiresIn || 60 * 5) * 1000),
 					});
 					const realBaseURL = new URL(ctx.context.baseURL);
@@ -352,13 +359,33 @@ export const magicLink = (options: MagicLinkOptions) => {
 						);
 						redirectWithError("EXPIRED_TOKEN");
 					}
-					await ctx.context.internalAdapter.deleteVerificationValue(
-						tokenValue.id,
-					);
-					const { email, name } = JSON.parse(tokenValue.value) as {
+					const {
+						email,
+						name,
+						attempt = 0,
+					} = JSON.parse(tokenValue.value) as {
 						email: string;
 						name?: string | undefined;
+            attempt: number;
 					};
+					if (attempt && attempt >= opts.allowedAttempts) {
+						await ctx.context.internalAdapter.deleteVerificationValue(
+							tokenValue.id,
+						);
+
+						throw ctx.redirect(`${errorCallbackURL}?error=INVALID_TOKEN`);
+					}
+					await ctx.context.internalAdapter.updateVerificationValue(
+						tokenValue.id,
+						{
+							value: JSON.stringify({
+								email,
+								name,
+								attempt: attempt + 1,
+							}),
+						},
+					);
+
 					let isNewUser = false;
 					let user = await ctx.context.internalAdapter
 						.findUserByEmail(email)
